requirements.txt
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.2
django-auth-ldap==2.2.0
django-filter==2.4.0
django-split-settings==1.0.1
django-suit==0.2.28
djangorestframework==3.12.2
drf-writable-nested==0.6.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.1
openapi-codec==1.3.2
packaging==20.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-ldap==3.3.1
pytz==2020.1
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
